+++ tests/assets/Settings8.php
@@ -199,6 +199,213 @@ class Settings8
                         'status' => 'RUNNING',
                         'type' => 'VISUAL_AB',
                     ),
+                3 =>
+                    array(
+                        'percentTraffic' => 100,
+                        'segments' => [],
+                        'goals' =>
+                            array(
+                                0 =>
+                                    array(
+                                        'identifier' => 'CUSTOME',
+                                        'id' => 409,
+                                        'type' => 'CUSTOM_GOAL',
+                                    ),
+                            ),
+                        'variations' =>
+                            array(
+                                0 =>
+                                    array(
+                                        'changes' =>
+                                            array(),
+                                        'id' => 1,
+                                        'variables' =>
+                                            array(
+                                                0 =>
+                                                    array(
+                                                        'value' => 10,
+                                                        'type' => 'integer',
+                                                        'key' => 'V1',
+                                                        'id' => 1,
+                                                    ),
+                                            ),
+                                        'isFeatureEnabled' => false,
+                                        'weight' => 50,
+                                        'name' => 'Control',
+                                    ),
+                                1 =>
+                                    array(
+                                        'changes' =>
+                                            array(),
+                                        'id' => 2,
+                                        'variables' =>
+                                            array(
+                                                0 =>
+                                                    array(
+                                                        'value' => 20,
+                                                        'type' => 'integer',
+                                                        'key' => 'V1',
+                                                        'id' => 1,
+                                                    ),
+                                            ),
+                                        'isFeatureEnabled' => false,
+                                        'weight' => 25,
+                                        'name' => 'Variation-1',
+                                    ),
+                                2 =>
+                                    array(
+                                        'changes' =>
+                                            array(),
+                                        'id' => 2,
+                                        'variables' =>
+                                            array(
+                                                0 =>
+                                                    array(
+                                                        'value' => 20,
+                                                        'type' => 'integer',
+                                                        'key' => 'V1',
+                                                        'id' => 1,
+                                                    ),
+                                            ),
+                                        'isFeatureEnabled' => false,
+                                        'weight' => 25,
+                                        'name' => 'Variation-1',
+                                    ),
+                            ),
+                        'id' => 92,
+                        'segments' =>
+                            array(),
+                        'key' => 'FEATURE_TEST_ALL_DISABLED',
+                        'status' => 'RUNNING',
+                        'type' => 'FEATURE_TEST',
+                    ),
+                4 =>
+                    array(
+                        'percentTraffic' => 0,
+                        'segments' => [],
+                        'goals' =>
+                            array(
+                                0 =>
+                                    array(
+                                        'identifier' => 'CUSTOME',
+                                        'id' => 409,
+                                        'type' => 'CUSTOM_GOAL',
+                                    ),
+                            ),
+                        'variations' =>
+                            array(
+                                0 =>
+                                    array(
+                                        'changes' =>
+                                            array(),
+                                        'id' => 1,
+                                        'variables' =>
+                                            array(
+                                                0 =>
+                                                    array(
+                                                        'value' => 10,
+                                                        'type' => 'integer',
+                                                        'key' => 'V1',
+                                                        'id' => 1,
+                                                    ),
+                                            ),
+                                        'isFeatureEnabled' => false,
+                                        'weight' => 50,
+                                        'name' => 'Control',
+                                    ),
+                                1 =>
+                                    array(
+                                        'changes' =>
+                                            array(),
+                                        'id' => 2,
+                                        'variables' =>
+                                            array(
+                                                0 =>
+                                                    array(
+                                                        'value' => 20,
+                                                        'type' => 'integer',
+                                                        'key' => 'V1',
+                                                        'id' => 1,
+                                                    ),
+                                            ),
+                                        'isFeatureEnabled' => false,
+                                        'weight' => 25,
+                                        'name' => 'Variation-1',
+                                    ),
+                                2 =>
+                                    array(
+                                        'changes' =>
+                                            array(),
+                                        'id' => 2,
+                                        'variables' =>
+                                            array(
+                                                0 =>
+                                                    array(
+                                                        'value' => 20,
+                                                        'type' => 'integer',
+                                                        'key' => 'V1',
+                                                        'id' => 1,
+                                                    ),
+                                            ),
+                                        'isFeatureEnabled' => false,
+                                        'weight' => 25,
+                                        'name' => 'Variation-1',
+                                    ),
+                            ),
+                        'id' => 92,
+                        'segments' =>
+                            array(),
+                        'key' => 'FEATURE_TEST_ZERO_TRAFFIC',
+                        'status' => 'RUNNING',
+                        'type' => 'FEATURE_TEST',
+                    ),
+                5 =>
+                    array(
+                        'id' => 91,
+                        'segments' =>
+                            array(),
+                        'status' => 'RUNNING',
+                        'percentTraffic' => 0,
+                        'type' => 'FEATURE_ROLLOUT',
+                        'segments' => [],
+                        'variables' =>
+                            array(
+                                0 =>
+                                    array(
+                                        'value' => 10,
+                                        'type' => 'integer',
+                                        'key' => 'V1',
+                                        'id' => 1,
+                                    ),
+                            ),
+                        'key' => 'FEATURE_ROLLOUT_ZERO_TRAFFIC',
+                        'variations' =>
+                            array(
+                                0 =>
+                                    array(
+                                        'id' => 1,
+                                        'name' => 'website',
+                                        'changes' =>
+                                            array(),
+                                        'weight' => 100,
+                                    ),
+                            ),
+                        'goals' =>
+                            array(
+                                0 =>
+                                    array(
+                                        'identifier' => 'h',
+                                        'id' => 209,
+                                        'type' => 'CUSTOM_GOAL',
+                                    ),
+                                1 =>
+                                    array(
+                                        'identifier' => 'jk',
+                                        'id' => 210,
+                                        'type' => 'CUSTOM_GOAL',
+                                    ),
+                            ),
+                    ),
             ),
         'accountId' => 0,
         'version' => 1,