+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.4] - 2020-05-14
+### Changed
+- Return `null` from APIs in case user is not part of campaign
+
 ## [1.7.3] - 2020-05-12
 ### Changed
 - Fixed and added new test cases for feature rollout and validation checks